# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk layanan reservasi booking hotel (IAE Tubes)",
        "title": "Layanan Reservasi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/bookings": {
            "post": {
                "description": "Endpoint untuk membuat reservasi awal dengan status LOCKED",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Membuat pesanan awal (mengunci kamar)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payload Reservasi",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Booking"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/addons": {
            "post": {
                "description": "Endpoint untuk menambahkan addon (sarapan, asuransi, dll) ke dalam booking yang sudah ada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Menambahkan layanan tambahan ke pesanan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Booking ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload Addon",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateBookingAddonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.BookingAddon"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}/summary": {
            "get": {
                "description": "Mendapatkan ringkasan rincian biaya kamar dan layanan tambahan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Menampilkan nota total (Summary)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Booking ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.BookingSummary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{id}/hold": {
            "post": {
                "description": "Endpoint untuk menahan kamar menggunakan Redis selama 10 menit agar tidak diambil orang lain saat mengisi form.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Menahan sementara kamar (Fase 1 Booking)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Room ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload Hold",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.HoldRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Endpoint untuk membebaskan kamar jika pengguna membatalkan isi form sebelum 10 menit.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Melepaskan tahanan kamar (Batal Booking)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Room ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload Hold (membutuhkan guest_id)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.HoldRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Booking": {
            "type": "object",
            "properties": {
                "addons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BookingAddon"
                    }
                },
                "check_in_date": {
                    "type": "string"
                },
                "check_out_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "grand_total": {
                    "type": "number"
                },
                "guest_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_addons_price": {
                    "type": "number"
                },
                "total_room_price": {
                    "type": "number"
                }
            }
        },
        "domain.BookingAddon": {
            "type": "object",
            "properties": {
                "addon_id": {
                    "type": "string"
                },
                "booking_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price_at_booking": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "domain.BookingSummary": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/domain.Booking"
                },
                "guest_details": {
                    "$ref": "#/definitions/domain.Guest"
                },
                "room_details": {
                    "$ref": "#/definitions/domain.Room"
                }
            }
        },
        "domain.CreateBookingAddonRequest": {
            "type": "object",
            "required": [
                "addon_id",
                "quantity"
            ],
            "properties": {
                "addon_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "domain.CreateBookingRequest": {
            "type": "object",
            "required": [
                "check_in_date",
                "check_out_date",
                "guest_id",
                "room_id"
            ],
            "properties": {
                "check_in_date": {
                    "description": "format YYYY-MM-DD",
                    "type": "string"
                },
                "check_out_date": {
                    "description": "format YYYY-MM-DD",
                    "type": "string"
                },
                "guest_id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Opsional: Detail error validasi"
                },
                "message": {
                    "description": "Detail pesan kesalahan",
                    "type": "string"
                },
                "status": {
                    "description": "Biasanya diisi \"error\" atau \"fail\"",
                    "type": "string"
                }
            }
        },
        "domain.Guest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ktp_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "domain.HoldRoomRequest": {
            "type": "object",
            "required": [
                "guest_id"
            ],
            "properties": {
                "guest_id": {
                    "type": "string"
                }
            }
        },
        "domain.Meta": {
            "type": "object",
            "properties": {
                "api_version": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "domain.Room": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "facilities": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_per_night": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Objek atau Array data utama"
                },
                "message": {
                    "description": "Pesan sukses",
                    "type": "string"
                },
                "meta": {
                    "description": "Informasi tambahan (opsional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Meta"
                        }
                    ]
                },
                "status": {
                    "description": "Biasanya diisi \"success\"",
                    "type": "string"
                }
            }
        }
    }
}